// Code/UICode/src/app/Interfaces/shipmentHistory.ts
export class ShipmentHistory {

    messageType: any;
    shipmentId: any;
    blockTimestamp: any;
     // deviceId: any;
    productList: any;
    fromUser: any;

    sourceLocation: any;
    destinationLocation: any;
    logisticsPartner: any;
    dateOfShipment: any;
    dateOfDelivery: any;
    purchaseOrderNumber: any;

    shipmentInfo: any;
    status: any;
    docName: any;
    docRef: any;
    docType: any;
    heartBeat: any;

    alertCount: any;
    alertDetails: any;

    address: any;
    blockHash: any;
    blockNumber: any;
    event: any;
    logIndex: any;
    removed: any;
    transactionHash: any;
    transactionIndex: any;
}